requirements.txt
openai==1.3.5
langchain==0.0.291
python-dotenv==1.0.0
faiss-cpu==1.7.4
pypdf==3.17.1